Add `Domainic::Type::Behavior::StringBehavior::MatchingBehavior`
Move string matching constraints into a dedicated module to improve code
organization and enable potential reuse. This refactoring isolates
matching functionality (equality, pattern matching, substring
inclusion/exclusion) from other string-specific constraints.

Changelog:
  * added `Domainic::Type::Behavior::StringBehavior::MatchingBehavior`
  * changed `Domainic::Type::Behavior::StringBehavior` to include
    MatchingBehavior

No changes to public interface or behavior * this is purely an internal
reorganization to improve maintainability and prepare for potential
reuse in other types that need matching capabilities.
