+++ domainic-type/lib/domainic/type/behavior/string_behavior.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require 'domainic/type/behavior/string_behavior/matching_behavior'
 require 'domainic/type/behavior/sizable_behavior'
 
 module Domainic
@@ -54,6 +55,7 @@ module Domainic
       # @author {https://aaronmallen.me Aaron Allen}
       # @since 0.1.0
       module StringBehavior
+        include MatchingBehavior
         include SizableBehavior
 
         # Validate string contains only alphanumeric characters.
@@ -101,24 +103,6 @@ module Domainic
           constrain :self, :emptiness, description: 'being'
         end
         alias empty being_empty
-
-        # Validate string equals a specific value.
-        #
-        # @example
-        #   type.being_equal_to("hello")
-        #   type.validate("hello")  # => true
-        #   type.validate("world")  # => false
-        #
-        # @param literal [String, Symbol] the value to compare against
-        # @return [Behavior] self for method chaining
-        # @rbs (String | Symbol literal) -> Behavior
-        def being_equal_to(literal)
-          # @type self: Object & Behavior
-          constrain :self, :equality, literal, description: 'equaling'
-        end
-        alias eql being_equal_to
-        alias equal_to being_equal_to
-        alias equaling being_equal_to
 
         # Validate string is lowercase.
         #
@@ -266,62 +250,6 @@ module Domainic
         alias not_being_lowercase being_uppercase
         alias uppercase being_uppercase
 
-        # Validate string contains all specified substrings.
-        #
-        # @example
-        #   type.containing("hello", "world")
-        #   type.validate("hello world")  # => true
-        #   type.validate("hello")        # => false
-        #
-        # @param literals [Array<String, Symbol>] the substrings to look for
-        # @return [Behavior] self for method chaining
-        # @rbs (*String | Symbol literals) -> Behavior
-        def containing(*literals)
-          # @type self: Object & Behavior
-          including = literals.map do |literal|
-            @constraints.prepare :self, :inclusion, literal, coerce_with: lambda(&:to_s)
-          end
-          constrain :self, :and, including, concerning: :inclusion
-        end
-        alias including containing
-
-        # Validate string does not contain any specified substrings.
-        #
-        # @example
-        #   type.excluding("foo", "bar")
-        #   type.validate("hello world")  # => true
-        #   type.validate("foo bar")      # => false
-        #
-        # @param literals [Array<String, Symbol>] the substrings to exclude
-        # @return [Behavior] self for method chaining
-        # @rbs (*String | Symbol literals) -> Behavior
-        def excluding(*literals)
-          # @type self: Object & Behavior
-          including = literals.map do |literal|
-            @constraints.prepare :self, :inclusion, literal, coerce_with: lambda(&:to_s)
-          end
-          constrain :self, :nor, including, concerning: :exclusion
-        end
-        alias omitting excluding
-
-        # Validate string matches all specified patterns.
-        #
-        # @example
-        #   type.matching(/^\w+$/, /\d/)
-        #   type.validate("hello123")  # => true
-        #   type.validate("hello")     # => false
-        #
-        # @param patterns [Array<String, Regexp>] the patterns to match against
-        # @return [Behavior] self for method chaining
-        # @rbs (*String | Regexp patterns) -> Behavior
-        def matching(*patterns)
-          # @type self: Object & Behavior
-          matching_patterns = patterns.map do |pattern|
-            @constraints.prepare :self, :match_pattern, pattern
-          end
-          constrain :self, :and, matching_patterns, concerning: :pattern_inclusion
-        end
-
         # Validate string is not empty.
         #
         # @example
@@ -335,43 +263,6 @@ module Domainic
           # @type self: Object & Behavior
           empty = @constraints.prepare :self, :emptiness
           constrain :self, :not, empty, concerning: :emptiness, description: 'being'
-        end
-
-        # Validate string does not equal a specific value.
-        #
-        # @example
-        #   type.not_being_equal_to("admin")
-        #   type.validate("user")   # => true
-        #   type.validate("admin")  # => false
-        #
-        # @param literal [String, Symbol] the value to compare against
-        # @return [Behavior] self for method chaining
-        # @rbs (String | Symbol literal) -> Behavior
-        def not_being_equal_to(literal)
-          # @type self: Object & Behavior
-          equal_to = @constraints.prepare :self, :equality, literal
-          constrain :self, :not, equal_to, concerning: :equality, description: 'being'
-        end
-        alias not_eql not_being_equal_to
-        alias not_equal_to not_being_equal_to
-        alias not_equaling not_being_equal_to
-
-        # Validate string does not match any specified patterns.
-        #
-        # @example
-        #   type.not_matching(/\d/, /[A-Z]/)
-        #   type.validate("hello")     # => true
-        #   type.validate("Hello123")  # => false
-        #
-        # @param patterns [Array<String, Regexp>] the patterns to avoid matching
-        # @return [Behavior] self for method chaining
-        # @rbs (*String | Regexp patterns) -> Behavior
-        def not_matching(*patterns)
-          # @type self: Object & Behavior
-          matching_patterns = patterns.map do |pattern|
-            @constraints.prepare :self, :match_pattern, pattern
-          end
-          constrain :self, :nor, matching_patterns, concerning: :pattern_exclusion
         end
       end
     end
