+++ domainic-type/sig/domainic/type/behavior/string_behavior.rbs
@@ -50,6 +50,8 @@ module Domainic
       # @author {https://aaronmallen.me Aaron Allen}
       # @since 0.1.0
       module StringBehavior
+        include MatchingBehavior
+
         include SizableBehavior
 
         # Validate string contains only alphanumeric characters.
@@ -89,23 +91,6 @@ module Domainic
         def being_empty: () -> Behavior
 
         alias empty being_empty
-
-        # Validate string equals a specific value.
-        #
-        # @example
-        #   type.being_equal_to("hello")
-        #   type.validate("hello")  # => true
-        #   type.validate("world")  # => false
-        #
-        # @param literal [String, Symbol] the value to compare against
-        # @return [Behavior] self for method chaining
-        def being_equal_to: (String | Symbol literal) -> Behavior
-
-        alias eql being_equal_to
-
-        alias equal_to being_equal_to
-
-        alias equaling being_equal_to
 
         # Validate string is lowercase.
         #
@@ -235,43 +220,6 @@ module Domainic
 
         alias uppercase being_uppercase
 
-        # Validate string contains all specified substrings.
-        #
-        # @example
-        #   type.containing("hello", "world")
-        #   type.validate("hello world")  # => true
-        #   type.validate("hello")        # => false
-        #
-        # @param literals [Array<String, Symbol>] the substrings to look for
-        # @return [Behavior] self for method chaining
-        def containing: (*String | Symbol literals) -> Behavior
-
-        alias including containing
-
-        # Validate string does not contain any specified substrings.
-        #
-        # @example
-        #   type.excluding("foo", "bar")
-        #   type.validate("hello world")  # => true
-        #   type.validate("foo bar")      # => false
-        #
-        # @param literals [Array<String, Symbol>] the substrings to exclude
-        # @return [Behavior] self for method chaining
-        def excluding: (*String | Symbol literals) -> Behavior
-
-        alias omitting excluding
-
-        # Validate string matches all specified patterns.
-        #
-        # @example
-        #   type.matching(/^\w+$/, /\d/)
-        #   type.validate("hello123")  # => true
-        #   type.validate("hello")     # => false
-        #
-        # @param patterns [Array<String, Regexp>] the patterns to match against
-        # @return [Behavior] self for method chaining
-        def matching: (*String | Regexp patterns) -> Behavior
-
         # Validate string is not empty.
         #
         # @example
@@ -281,34 +229,6 @@ module Domainic
         #
         # @return [Behavior] self for method chaining
         def not_being_empty: () -> Behavior
-
-        # Validate string does not equal a specific value.
-        #
-        # @example
-        #   type.not_being_equal_to("admin")
-        #   type.validate("user")   # => true
-        #   type.validate("admin")  # => false
-        #
-        # @param literal [String, Symbol] the value to compare against
-        # @return [Behavior] self for method chaining
-        def not_being_equal_to: (String | Symbol literal) -> Behavior
-
-        alias not_eql not_being_equal_to
-
-        alias not_equal_to not_being_equal_to
-
-        alias not_equaling not_being_equal_to
-
-        # Validate string does not match any specified patterns.
-        #
-        # @example
-        #   type.not_matching(/\d/, /[A-Z]/)
-        #   type.validate("hello")     # => true
-        #   type.validate("Hello123")  # => false
-        #
-        # @param patterns [Array<String, Regexp>] the patterns to avoid matching
-        # @return [Behavior] self for method chaining
-        def not_matching: (*String | Regexp patterns) -> Behavior
       end
     end
   end
