+++ domainic-type/sig/domainic/type/behavior/string_behavior/matching_behavior.rbs
@@ -1,0 +1,84 @@
+module Domainic
+  module Type
+    module Behavior
+      module StringBehavior
+        # A module providing string matching constraint methods
+        #
+        # This module extends types with methods for constraining string values based on
+        # equality, pattern matching, and substring inclusion/exclusion. It provides a
+        # fluent interface for building complex string matching constraints.
+        #
+        # @example Basic equality constraints
+        #   type = StringType.new
+        #     .being_equal_to("expected")
+        #     .not_being_equal_to("forbidden")
+        #
+        # @example Pattern matching constraints
+        #   type = StringType.new
+        #     .matching(/^\w+$/, /[0-9]/)
+        #     .not_matching(/admin/i)
+        #
+        # @example Substring constraints
+        #   type = StringType.new
+        #     .containing("allowed", "required")
+        #     .excluding("forbidden", "blocked")
+        #
+        # @author {https://aaronmallen.me Aaron Allen}
+        # @since 0.1.0
+        module MatchingBehavior
+          # Constrain string to equal a specific value
+          #
+          # @param literal [String, Symbol] the value to match against
+          # @return [Behavior] self for method chaining
+          def being_equal_to: (String | Symbol literal) -> Behavior
+
+          alias eql being_equal_to
+
+          alias equal_to being_equal_to
+
+          alias equaling being_equal_to
+
+          # Constrain string to contain specific substrings
+          #
+          # @param literals [Array<String, Symbol>] the required substrings
+          # @return [Behavior] self for method chaining
+          def containing: (*String | Symbol literals) -> Behavior
+
+          alias including containing
+
+          # Constrain string to exclude specific substrings
+          #
+          # @param literals [Array<String, Symbol>] the forbidden substrings
+          # @return [Behavior] self for method chaining
+          def excluding: (*String | Symbol literals) -> Behavior
+
+          alias omitting excluding
+
+          # Constrain string to match specific patterns
+          #
+          # @param patterns [Array<String, Regexp>] the required patterns
+          # @return [Behavior] self for method chaining
+          def matching: (*String | Regexp patterns) -> Behavior
+
+          # Constrain string to not equal a specific value
+          #
+          # @param literal [String, Symbol] the forbidden value
+          # @return [Behavior] self for method chaining
+          def not_being_equal_to: (String | Symbol literal) -> Behavior
+
+          alias not_eql not_being_equal_to
+
+          alias not_equal_to not_being_equal_to
+
+          alias not_equaling not_being_equal_to
+
+          # Constrain string to not match specific patterns
+          #
+          # @param patterns [Array<String, Regexp>] the forbidden patterns
+          # @return [Behavior] self for method chaining
+          def not_matching: (*String | Regexp patterns) -> Behavior
+        end
+      end
+    end
+  end
+end
