+++ domainic-type/spec/domainic/type/behavior/string_behavior_spec.rb
@@ -44,15 +44,15 @@ RSpec.describe Domainic::Type::Behavior::StringBehavior do
     end
   end
 
-  describe '#being_equal_to' do
-    context 'when validating a matching string' do
-      subject(:validation) { type.being_equal_to('hello').validate('hello') }
+  describe '#being_empty' do
+    context 'when validating an empty string' do
+      subject(:validation) { type.being_empty.validate('') }
 
       it { is_expected.to be true }
     end
 
-    context 'when validating a non-matching string' do
-      subject(:validation) { type.being_equal_to('hello').validate('world') }
+    context 'when validating a non-empty string' do
+      subject(:validation) { type.being_empty.validate('not empty') }
 
       it { is_expected.to be false }
     end
@@ -170,57 +170,15 @@ RSpec.describe Domainic::Type::Behavior::StringBehavior do
     end
   end
 
-  describe '#containing' do
-    context 'when validating a string containing all specified substrings' do
-      subject(:validation) { type.containing('hello', 'world').validate('hello world') }
+  describe '#not_being_empty' do
+    context 'when validating a non-empty string' do
+      subject(:validation) { type.not_being_empty.validate('not empty') }
 
       it { is_expected.to be true }
     end
 
-    context 'when validating a string missing specified substrings' do
-      subject(:validation) { type.containing('hello', 'there').validate('hello world') }
-
-      it { is_expected.to be false }
-    end
-  end
-
-  describe '#excluding' do
-    context 'when validating a string without specified substrings' do
-      subject(:validation) { type.excluding('foo', 'bar').validate('hello world') }
-
-      it { is_expected.to be true }
-    end
-
-    context 'when validating a string with specified substrings' do
-      subject(:validation) { type.excluding('hello', 'world').validate('hello world') }
-
-      it { is_expected.to be false }
-    end
-  end
-
-  describe '#matching' do
-    context 'when validating a string matching all patterns' do
-      subject(:validation) { type.matching(/^\w+$/, /\d/).validate('hello123') }
-
-      it { is_expected.to be true }
-    end
-
-    context 'when validating a string not matching all patterns' do
-      subject(:validation) { type.matching(/^\w+$/, /\d/).validate('hello') }
-
-      it { is_expected.to be false }
-    end
-  end
-
-  describe '#not_matching' do
-    context 'when validating a string not matching any patterns' do
-      subject(:validation) { type.not_matching(/\d/, /[A-Z]/).validate('hello') }
-
-      it { is_expected.to be true }
-    end
-
-    context 'when validating a string matching some patterns' do
-      subject(:validation) { type.not_matching(/\d/, /[A-Z]/).validate('hello123') }
+    context 'when validating an empty string' do
+      subject(:validation) { type.not_being_empty.validate('') }
 
       it { is_expected.to be false }
     end
